langgraph-agent/requirements.txt
flask
langgraph
langchain
langchain-community
langchain-groq
python-dotenv
duckduckgo-search
